Amelioration du moteur + accepte les phrases
