+++ db/migrate/20220604111323_add_lyricsto_game_session_songs.rb
@@ -1,0 +1,5 @@
+class AddLyricstoGameSessionSongs < ActiveRecord::Migration[6.1]
+  def change
+    add_column :game_session_songs, :guessed_splitted_lyrics, :jsonb, default: []
+  end
+end
